+++ pylatex/table.py
@@ -10,7 +10,8 @@ from .base_classes import LatexObject, Container, Command, UnsafeCommand, \
     Float, Environment
 from .package import Package
 from .errors import TableRowSizeError, TableError
-from .utils import dumps_list, NoEscape, escape_latex
+from .utils import dumps_list, NoEscape
+import pylatex.config as cf
 
 from collections import Counter
 import re
@@ -54,8 +55,8 @@ class Tabular(Environment):
     }
 
     def __init__(self, table_spec, data=None, pos=None, *,
-                 row_height=None, col_space=None, width=None, arguments=None,
-                 **kwargs):
+                 row_height=None, col_space=None, width=None, booktabs=None,
+                 arguments=None, **kwargs):
         """
         Args
         ----
@@ -68,13 +69,18 @@ class Tabular(Environment):
             row height
         col_space: str
             Specifies the spacing between table columns
-        arguments: str or `list`
-            The arguments to append to the table
+        booktabs: bool
+            Enable or disable booktabs style tables. These tables generally
+            look nicer than regular tables. If this is `None` it will use the
+            value of the ``booktabs`` attribte from the `~.active`
+            configuration. This attribute is `False` by default.
         width: int
             The amount of columns that the table has. If this is `None` it is
             calculated based on the ``table_spec``, but this is only works for
             simple specs. In cases where this calculation is wrong override the
             width using this argument.
+        arguments: str or `list`
+            The arguments to append to the table
 
         References
         ----------
@@ -85,6 +91,13 @@ class Tabular(Environment):
             self.width = _get_table_width(table_spec)
         else:
             self.width = width
+
+        if booktabs is None:
+            booktabs = cf.active.booktabs
+        self.booktabs = booktabs
+
+        if self.booktabs:
+            self.packages.add(Package('booktabs'))
 
         self.row_height = row_height
         self.col_space = col_space
@@ -123,8 +136,36 @@ class Tabular(Environment):
 
         return string + super().dumps()
 
-    def add_hline(self, start=None, end=None, *, color=None):
-        """Add a horizontal line to the table.
+    def dumps_content(self, **kwargs):
+        r"""Represent the content of the tabular in LaTeX syntax.
+
+        This adds the top and bottomrule when using a booktabs style tabular.
+
+        Args
+        ----
+        \*\*kwargs:
+            Arguments that can be passed to `~.dumps_list`
+
+        Returns
+        -------
+        string:
+            A LaTeX string representing the
+        """
+
+        content = ''
+        if self.booktabs:
+            content += '\\toprule%\n'
+
+        content += super().dumps_content(**kwargs)
+
+        if self.booktabs:
+            content += '\\bottomrule%\n'
+
+        return NoEscape(content)
+
+    def add_hline(self, start=None, end=None, *, color=None,
+                  cmidruleoption=None):
+        r"""Add a horizontal line to the table.
 
         Args
         ----
@@ -134,10 +175,20 @@ class Tabular(Environment):
             At what cell the line should end
         color: str
             The hline color.
+        cmidruleoption: str
+            The option to be used for the booktabs cmidrule, i.e. the ``x`` in
+            ``\cmidrule(x){1-3}``.
         """
+        if self.booktabs:
+            hline = 'midrule'
+            cline = 'cmidrule'
+            if cmidruleoption is not None:
+                cline += '(' + cmidruleoption + ')'
+        else:
+            hline = 'hline'
+            cline = 'cline'
 
         if color is not None:
-            # TODO: This check should not be needed.
             if not self.color:
                 self.packages.append(Package('xcolor', options='table'))
                 self.color = True
@@ -145,18 +196,15 @@ class Tabular(Environment):
             self.append(color_command)
 
         if start is None and end is None:
-            self.append(NoEscape(r'\hline'))
+            self.append(Command(hline))
         else:
             if start is None:
                 start = 1
             elif end is None:
                 end = self.width
 
-            if self.escape:
-                start = escape_latex(start)
-                end = escape_latex(end)
-
-            self.append(UnsafeCommand('cline', start + '-' + end))
+            self.append(Command(cline,
+                                dumps_list([start, NoEscape('-'), end])))
 
     def add_empty_row(self):
         """Add an empty row to the table."""