+++ docs/source/changelog.rst
@@ -75,6 +75,7 @@ Added
   based on the ``table_spec`` argument.
 - Default configuration for certain options can be overwritten with the new
   `pylatex.config` module.
+- Add support for booktabs tables, which look nicer than normal tables.
 
 Fixed
 ~~~~~